+++ Emergence/Models/Electric Objects/Artwork.swift
@@ -7,6 +7,7 @@ struct Artwork: Artworkable {
     let medium: String
     
     let artists: [Artistable]?
+    let culturalMarker: String?
     let images: [Imageable]
 
     lazy var defaultImage: Imageable? = {
@@ -16,7 +17,13 @@ struct Artwork: Artworkable {
 
         return defaultImages.isNotEmpty ? defaultImages.first : self.images.first
     }()
-    
+
+    func oneLinerArtist() -> String? {
+        if let artist = artists?.first {
+            return artist.name
+        }
+        return culturalMarker
+    }
 }
 
 extension Artwork: Decodable {
@@ -34,7 +41,14 @@ extension Artwork: Decodable {
         title = titleValue
         medium = mediumValue
 
-        artists = []
+        culturalMarker = "cultural_marker" <~~ json
+
+        if let artistsValue: [Artist] = "artists" <~~ json {
+            artists = artistsValue.map { return $0 as Artistable }
+        } else {
+            artists = []
+        }
+
         if let imagesValue: [Image] = "images" <~~ json {
            images = imagesValue.map { return $0 as Imageable }
         } else {
